+++ component/pyroscope/scrape/internal/fastdelta/location_index.go
