+++ component/pyroscope/scrape/internal/fastdelta/location_index_test.go
