+++ component/pyroscope/scrape/internal/pproflite/pproflite_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/google/pprof/profile"
 	"github.com/stretchr/testify/require"
 
-	"github.com/grafana/agent/component/pyroscope/scrape/pproflite"
+	"github.com/grafana/agent/component/pyroscope/scrape/internal/pproflite"
 )
 
 func TestDecoderEncoder(t *testing.T) {
